feat: 설정 페이지 추가 및 레이아웃 컴포넌트 수정
