+++ src/app/(private)/setting/family/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <div>가족 설정 페이지</div>;
+}
