+++ src/app/(private)/setting/user/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <div>유저 설정 페이지</div>;
+}
