+++ src/app/(public)/signup/page.tsx
@@ -6,7 +6,8 @@ import { useMutation } from "@tanstack/react-query";
 import clsx from "clsx";
 import { FormProvider, useForm } from "react-hook-form";
 import { authApi } from "@/api";
-import { HeaderLayout, Spinner, customToast } from "@/components";
+import { Spinner, customToast } from "@/components";
+import { HeaderLayout } from "@/components/ui";
 import { SignupFormType, SignupSchema } from "@/schemas/SignupSchema";
 import PasswordInput from "./PasswordInput";
 import SignupAgreements from "./SignupAgreements";
