+++ src/components/index.ts
@@ -3,7 +3,6 @@ export { default as CheckBox } from "./CheckBox";
 export { default as CheckList } from "./CheckList";
 export { default as ColorSelect } from "./ColorSelect";
 export { default as CustomCharacter } from "./CustomCharacter";
-export { default as HeaderLayout } from "./HeaderLayout";
 export { default as LargeInput } from "./LargeInput";
 export { default as LinkButton } from "./LinkButton";
 export { default as Triangle } from "./Triangle";
