+++ src/components/ui/HeaderLayout/index.ts
@@ -1,0 +1,8 @@
+import Content from "./Content";
+import Progressbar from "./Progressbar";
+import Title from "./Title";
+import Wrapper from "./Wrapper";
+
+const HeaderLayout = { Wrapper, Progressbar, Title, Content };
+
+export default HeaderLayout;
